Create README.md
MEME-SHARE APP
Giving a rough layout of the app how it going to look like 
https://whimsical.com/meme-share-app-3BkbuLUQBEW4K1sKLQZZ6s please go through this link viewers as you will get a rough idea of the app
Technology which is used are:  1) ANDROID STUDIO
				2)KOTLIN LANGUAGE

API link: https://meme-api.herokuapp.com/gimme.

Description 
It’s a beginner-level Meme share app. It contains two buttons one is a Share Button and one is the Next Button. The share button is used to share the meme in different sharing functionally platforms e.g Gmail, Whatsapp, Instagram, clipboard, etc. The next button is used for calling the next meme from the API. There is also a circular loader that will appear when the next button is clicked and the screen is about to load the next meme. The background of the button is purple and The background of the screen is black.
Tip: while doing android projects in Kotlin you must sync an (id: Kotlin-android-extensions) otherwise it will show a reference error.